misc
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@289 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ ThickSlabVR.mm
@@ -249,7 +249,7 @@ extern short Altivec;
 	}
 }
 
--(void) setWLWW: (long) l :(long) w
+-(void) setWLWW: (float) l :(float) w
 {
 	wl = l;
 	ww = w;
@@ -273,7 +273,7 @@ extern short Altivec;
 	}
 }
 
--(void) setBlendingWLWW: (long) l :(long) w
+-(void) setBlendingWLWW: (float) l :(float) w
 {
 	if( imageBlendingPtr)
 	{
